JBRULES-2835: revert ksession id from long to int

+++ kogito-api/drools-api/src/main/java/org/drools/persistence/jpa/JPAKnowledgeService.java
@@ -111,7 +111,7 @@ public class JPAKnowledgeService {
                                                                              environment );
     }
 
-    public static StatefulKnowledgeSession loadStatefulKnowledgeSession(long id,
+    public static StatefulKnowledgeSession loadStatefulKnowledgeSession(int id,
                                                                         KnowledgeBase kbase,
                                                                         KnowledgeSessionConfiguration configuration,
                                                                         Environment environment) {
